% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_legend.R
\name{add_legend}
\alias{add_legend}
\title{Customized Legend for CE and RP plots.}
\usage{
add_legend(...)
}
\arguments{
\item{...}{free}
}
\description{
Customized Legend for CE and RP plots.
}
