% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_en.R
\name{ce_en}
\alias{ce_en}
\title{Certainty Equivalent Function for Negative Exponential Function}
\usage{
ce_en(profit, rac, weight = 0)
}
\arguments{
\item{profit}{\code{data.frame} with profit values}

\item{rac}{scalar of RAC value}

\item{weight}{original wealth}
}
\value{
Scalar with the CE value
}
\description{
Certainty Equivalent Function for Negative Exponential Function
}
