% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_p.R
\name{ce_p}
\alias{ce_p}
\title{Certainty Equivalent Function with Power Utility Function}
\usage{
ce_p(profit, rac, weight = 0)
}
\arguments{
\item{profit}{\code{data.frame} with profit values}

\item{rac}{scalar of RAC value}

\item{weight}{original wealth}
}
\value{
Scalar with the CE value
}
\description{
Certainty Equivalent Function with Power Utility Function
}
