% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_power.R
\name{ce_power}
\alias{ce_power}
\title{Certainty Equivalent Computation using Power Utility Function}
\usage{
ce_power(data, rac_ini, rac_fin, weight = 0)
}
\arguments{
\item{data}{\code{data.frame} with profit values}

\item{rac_ini}{Initial value for the RAC sequence}

\item{rac_fin}{Final value for the RAC sequence}

\item{weight}{Original wealth}
}
\value{
Generate three objects: A table with the CEs, a vector
of risk aversion coefficients RAC, and a plot to compare the CEs.
}
\description{
Certainty Equivalent Computation using Power Utility Function
}
