% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ceRtainty.R
\name{plot.ceRtainty}
\alias{plot.ceRtainty}
\title{Plot class function}
\usage{
\method{plot}{ceRtainty}(CER)
}
\arguments{
\item{data}{...}

\item{rac}{...}

\item{rac_ini}{...}

\item{rac_fin}{...}

\item{rac_len}{...}
}
\description{
Plot class function
}
