% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ce_en.R
\name{plot_ce_en}
\alias{plot_ce_en}
\title{Plot for CE using Exponential Negative Utility Function}
\usage{
plot_ce_en(data, rac = 0, rac_ini = 0, rac_fin = 1, rac_len = 10)
}
\arguments{
\item{data}{Data set with CE already computed}

\item{rac}{Scalar with the RAC to use in the CE computation.
When the analysis consider only one value of RAC}

\item{rac_ini}{Vector of the RAC to use in the CE computation.
When the analysis consider a sequence of RAC values}

\item{rac_fin}{Final value for the RAC vector}

\item{rac_len}{RAC vector length}
}
\value{
Plot of CE to compare treatments/projects
}
\description{
Plot for CE using Exponential Negative Utility Function
}
