% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk_premium_en.R
\name{plot_risk_premium_en}
\alias{plot_risk_premium_en}
\title{Plot of the Risk Premium values using Exponential Negative Utility Function}
\usage{
plot_risk_premium_en(data, rac_ini, rac_fin, rac_len)
}
\arguments{
\item{data}{\code{data.frame} of CE computed by Exponential Negative function}

\item{rac_ini}{Initial RAC values used in the CE computation}

\item{rac_fin}{Final RAC values used in the CE computation}

\item{rac_len}{Length of the RAC vector used in the CE computation}
}
\value{
Plot object
}
\description{
Plot of the Risk Premium values using Exponential Negative Utility Function
}
