% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profitSWG-data.R
\docType{data}
\name{profitSWG}
\alias{profitSWG}
\title{Profit SWG}
\format{A data frame with 8 rows of profit and 4 pesticide treatments:
\describe{
  \item{control}{Control (non-treated case), in US dollars}
  \item{fracture}{Fracture treatment, in US dollars}
  \item{milstop}{Milstop treatment, in US dollars}
  \item{serenade}{Serenade Optimum treatment, in US dollars}
  }}
\source{
Soto-Caro, Wu, Guan (2019). "Evaluating Pest Management Strategies:
A Robust Method and Its Application to Strawberry Disease Management".
AAEA 2019 Conference.
}
\usage{
data(profitSWG)
}
\description{
The data come from strawberry trials experiments in Florida, USA, performed
by the Gulf Coast Research and Education Center, University of Florida.
}
\details{
Correspond to the profit for three pesticide treatments plus the case without
treatment. Four observations for each season, 2014-15 and 2015-16.
}
\examples{
data(profitSWG)
summary(profitSWG)
}
\references{
Soto-Caro, Wu, Guan (2019). "Evaluating Pest Management Strategies:
A Robust Method and Its Application to Strawberry Disease Management".
AAEA 2019 Conference.
}
\keyword{datasets}
