% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_len.R
\name{rac_len}
\alias{rac_len}
\title{Define the length of the Risk Aversion Coefficient, RAC.}
\usage{
rac_len(ini, fin, data)
}
\arguments{
\item{ini}{The initial value of the RAC sequence}

\item{fin}{The final value of the RAC sequence}

\item{data}{Original data, could be a vector or a matrix}
}
\value{
Two elements are generated: "r" is the RAC vector, and "length" is
a scalar with the number of elements on RAC vector.
}
\description{
Define the length of the Risk Aversion Coefficient, RAC.
}
