% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rac_seq.R
\name{rac_seq}
\alias{rac_seq}
\title{RAC Sequence generator}
\usage{
rac_seq(ini, fin, len)
}
\arguments{
\item{ini}{The initial value for the RAC}

\item{fin}{The final value for the RAC}

\item{len}{The Length of the vector to creates}
}
\value{
Vector of RACs
}
\description{
RAC Sequence generator
}
