% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACa.R
\name{RACa}
\alias{RACa}
\title{Adjusted Risk Aversion Coefficient (RACa)}
\usage{
RACa(rac, data)
}
\arguments{
\item{rac}{An scalar with the value of the relative RAC}

\item{data}{Dataset to weight the RAC}
}
\value{
This function create an adjustment to the relative risk
aversion coefficient, following Hardaker et al (2004).
}
\description{
Adjusted Risk Aversion Coefficient (RACa)
}
